requirements.txt
# Core dependencies
requests>=2.31.0
rich>=13.7.0

# Data processing
pandas>=2.1.0
sqlalchemy>=2.0.0
h5py>=3.10.0  # For HDF5 storage

# Text analysis
textstat>=0.7.3

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.22.0

# Optional but recommended
python-dotenv>=1.0.0  # For environment variable management
tqdm>=4.66.0         # For progress bars 

# App
streamlit>=1.29.0

# XML Processing
lxml>=4.9.0

# Embeddings and Vector Search
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4  # Explicitly using CPU version
pinecone-client>=2.2.0  # For vector database service
